+++ src/components/sidebar/tabs/NodeLibrarySideBarTab.vue
@@ -1,5 +1,5 @@
 <template>
-  <SideBarHeader :title="$t('sideToolBar.nodeLibrary')">
+  <SideBarTabTemplate :title="$t('sideToolBar.nodeLibrary')">
     <template #tool-buttons>
       <ToggleButton
         v-model:model-value="alphabeticalSort"
@@ -13,66 +13,68 @@
       >
       </ToggleButton>
     </template>
-  </SideBarHeader>
-  <TreePlus
-    class="node-lib-tree"
-    v-model:expandedKeys="expandedKeys"
-    selectionMode="single"
-    :value="renderedRoot.children"
-    :filter="true"
-    filterMode="lenient"
-    dragSelector=".p-tree-node-leaf"
-    :pt="{
-      nodeLabel: 'node-lib-tree-node-label',
-      nodeChildren: ({ props }) => ({
-        'data-comfy-node-name': props.node?.data?.name,
-        onMouseenter: (event: MouseEvent) => {
-          hoveredComfyNodeName = props.node?.data?.name
+    <template #body>
+      <TreePlus
+        class="node-lib-tree"
+        v-model:expandedKeys="expandedKeys"
+        selectionMode="single"
+        :value="renderedRoot.children"
+        :filter="true"
+        filterMode="lenient"
+        dragSelector=".p-tree-node-leaf"
+        :pt="{
+          nodeLabel: 'node-lib-tree-node-label',
+          nodeChildren: ({ props }) => ({
+            'data-comfy-node-name': props.node?.data?.name,
+            onMouseenter: (event: MouseEvent) => {
+              hoveredComfyNodeName = props.node?.data?.name
 
-          const hoverTarget = event.target as HTMLElement
-          const targetRect = hoverTarget.getBoundingClientRect()
-          nodePreviewStyle.top = `${targetRect.top - 40}px`
-          nodePreviewStyle.left = `${targetRect.right}px`
-        },
-        onMouseleave: () => {
-          hoveredComfyNodeName = null
-        }
-      })
-    }"
-  >
-    <template #folder="{ node }">
-      <span class="folder-label">{{ node.label }}</span>
-      <Badge
-        :value="node.totalNodes"
-        severity="secondary"
-        :style="{ marginLeft: '0.5rem' }"
-      ></Badge>
+              const hoverTarget = event.target as HTMLElement
+              const targetRect = hoverTarget.getBoundingClientRect()
+              nodePreviewStyle.top = `${targetRect.top - 40}px`
+              nodePreviewStyle.left = `${targetRect.right}px`
+            },
+            onMouseleave: () => {
+              hoveredComfyNodeName = null
+            }
+          })
+        }"
+      >
+        <template #folder="{ node }">
+          <span class="folder-label">{{ node.label }}</span>
+          <Badge
+            :value="node.totalNodes"
+            severity="secondary"
+            :style="{ marginLeft: '0.5rem' }"
+          ></Badge>
+        </template>
+        <template #node="{ node }">
+          <span class="node-label">{{ node.label }}</span>
+        </template>
+      </TreePlus>
+      <div
+        v-if="hoveredComfyNode"
+        class="node-lib-node-preview"
+        :style="nodePreviewStyle"
+      >
+        <NodePreview
+          :key="hoveredComfyNode.name"
+          :nodeDef="hoveredComfyNode"
+        ></NodePreview>
+      </div>
     </template>
-    <template #node="{ node }">
-      <span class="node-label">{{ node.label }}</span>
-    </template>
-  </TreePlus>
-  <div
-    v-if="hoveredComfyNode"
-    class="node-lib-node-preview"
-    :style="nodePreviewStyle"
-  >
-    <NodePreview
-      :key="hoveredComfyNode.name"
-      :nodeDef="hoveredComfyNode"
-    ></NodePreview>
-  </div>
+  </SideBarTabTemplate>
 </template>
 
 <script setup lang="ts">
 import Badge from 'primevue/badge'
 import ToggleButton from 'primevue/togglebutton'
-import SideBarHeader from '@/components/sidebar/SideBarHeader.vue'
 import { ComfyNodeDefImpl, useNodeDefStore } from '@/stores/nodeDefStore'
 import { computed, ref } from 'vue'
 import { TreeNode } from 'primevue/treenode'
 import TreePlus from '@/components/primevueOverride/TreePlus.vue'
 import NodePreview from '@/components/NodePreview.vue'
+import SideBarTabTemplate from '@/components/sidebar/tabs/SideBarTabTemplate.vue'
 
 const nodeDefStore = useNodeDefStore()
 const alphabeticalSort = ref(false)
